feat: rename forgot to recover
