+++ src/features/auth/RegisterForm/RegisterForm.styles.ts
@@ -21,6 +21,6 @@ export const FieldContainer = styled("div", {
   rowGap: "$10",
 });
 
-export const ForgotPasswordLink = styled(Typography.Link, {
+export const RecoverPasswordLink = styled(Typography.Link, {
   marginInline: "auto",
 });
