+++ src/features/password/RecoverPasswordForm/index.ts
@@ -1,0 +1,1 @@
+export * from "./RecoverPasswordForm";
