+++ src/features/password/index.ts
@@ -1,3 +1,3 @@
-export * from "./ForgotPasswordForm";
+export * from "./RecoverPasswordForm";
 export * from "./ResetPasswordForm";
 export * from "./ResetPasswordSuccess";
